make obtain validator common
